-- fix
		1. 增加对BK芯片的复位，暂可读取芯片的ID
		2. 修改BK芯片读取数据时的大小序，采用MSB

+++ develop_lib/hardware/bk953x/bk953x.h
@@ -21,7 +21,7 @@ typedef enum
 typedef struct 
 {
     i2c_type_e i2c_type;
-    uint16_t   chip_id;
+    uint32_t   chip_id;
     bk953x_band_type_e band_type;
 } bk953x_object_t;
